% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{o2_model_lsoda_cpp}
\alias{o2_model_lsoda_cpp}
\title{Patankar scheme for modelling oxygen depletion}
\usage{
o2_model_lsoda_cpp(
  yini,
  times,
  forcings,
  events,
  Flux,
  Khalf,
  Theta,
  iterations
)
}
\arguments{
\item{times}{An integer vector}

\item{Flux}{A numeric vector of O2 fluxes. One value per iteration}

\item{Khalf}{A numeric vector for Khalf. One value per iteration}

\item{Theta}{A numeric vector for Theta. One value per iteration}

\item{y}{start concentration of O2}
}
\description{
This function returns a data frame containing results from the oxygen model.
}
